Small Changes to Control

+++ LCU_MASTER/Core/Inc/LCU_Control/LevitationPositionCalculator.hpp
@@ -4,16 +4,16 @@
 #include "math.h"
 #include "LevitationPosition.hpp"
 
-class LevitationPositionCalculator : public ControlBlock<double(*)[8], LevitationPosition>{
+class LevitationPositionCalculator : public ControlBlock<float(*)[8], LevitationPosition>{
 public:
     static constexpr double x_sragy = 0.8069, y_sragy = 0.1042, z_sragy = 0.10825; 
     static constexpr double x_sragz = 0.81231, y_sragz = 0.2114, z_sragz = -0.0396;
     static constexpr double z_infr = 0.6328;
     static constexpr double grosorz = 0.015, grosory = 0.005;
 
-    double(&distances)[8] = *this->input_value;
+    float(&distances)[8] = *this->input_value;
 
-    LevitationPositionCalculator(double(&distances)[8]){
+    LevitationPositionCalculator(float(&distances)[8]){
         this->input_value = &distances;
         output_value = {0.0,0.0,0.0,0.0,0.0};
     };
